requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SocketIO==4.0.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
osmapi==1.2.2
PyJWT==1.7.1
pyroutelib3==1.1
pytest==5.0.0
python-dateutil==2.8.0
python-engineio==3.5.2
python-socketio==4.0.2
requests==2.21.0
six==1.12.0
urllib3==1.24.2
waitress==1.3.0
Werkzeug==0.15.2